Revert "Post to channel"
