+++ src/index.js
@@ -4,7 +4,8 @@ const app = require('express')();
 const bodyParser = require('body-parser');
 const initalEvent = require('./app/routes/events/initial');
 const slashWelcome = require('./app/routes/slash/welcome');
-const initalResponse = require('./app/routes/interactive/initialResponse');
+const initalResponse = require('./app/routes/interactive/initialRefactor');
+
 
 //const app = express();
